Revert "fix(codex): manage Orca worker thread lifecycle"
This reverts commit d10f9e30b3d5fd8ef556e6fa4b4798e4fb22ac40.
