+++ src/main/codex/codex-session-backfill.test.ts
@@ -1,7 +1,9 @@
+/* eslint-disable max-lines -- Why: backfill failure, race, archive, and marker scenarios share one hoisted filesystem mock whose state must reset uniformly. */
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import {
   appendFileSync,
   existsSync,
+  linkSync,
   lstatSync,
   mkdirSync,
   mkdtempSync,
@@ -190,6 +192,10 @@ function getSystemSessionsRoot(): string {
   return join(fakeHomeDir, '.codex', 'sessions')
 }
 
+function getSystemArchivedSessionsRoot(): string {
+  return join(fakeHomeDir, '.codex', 'archived_sessions')
+}
+
 function getManagedSessionsRoot(): string {
   return join(userDataDir, 'codex-runtime-home', 'home', 'sessions')
 }
@@ -267,6 +273,236 @@ afterEach(() => {
 })
 
 describe('backfillManagedCodexSessionsIntoSystemHome', () => {
+  it('hardlinks managed rollout files into the real home preserving layout', async () => {
+    const managedPath = writeManagedSession(
+      join('2026', '05', '26', 'rollout-a.jsonl'),
+      '{"type":"session_meta","id":"a"}\n'
+    )
+    writeManagedSession(join('2026', '06', '01', 'rollout-b.jsonl'), '{"id":"b"}\n')
+    writeFileSync(join(getManagedSessionsRoot(), '2026', '05', '26', 'notes.txt'), 'skip me\n')
+
+    const summary = await backfillManagedCodexSessionsIntoSystemHome(
+      resolveCodexSessionBackfillPaths()
+    )
+
+    expect(summary).toMatchObject({ scannedFiles: 2, linkedFiles: 2, failedFiles: 0 })
+    const targetPath = join(getSystemSessionsRoot(), '2026', '05', '26', 'rollout-a.jsonl')
+    expect(lstatSync(targetPath).ino).toBe(lstatSync(managedPath).ino)
+    expect(existsSync(join(getSystemSessionsRoot(), '2026', '06', '01', 'rollout-b.jsonl'))).toBe(
+      true
+    )
+    expect(existsSync(join(getSystemSessionsRoot(), '2026', '05', '26', 'notes.txt'))).toBe(false)
+    expect(readAuditActions()).toEqual(['hardlink', 'hardlink', 'run-summary'])
+  })
+
+  it('only backfills rollout files in the exact YYYY/MM/DD layout', async () => {
+    writeManagedSession(join('2026', '05', '26', 'rollout-valid ü.jsonl'), 'valid\n')
+    writeManagedSession(join('2026', '05', '26', 'session-index.jsonl'), 'not a rollout\n')
+    writeManagedSession(join('2026', '5', '26', 'rollout-wrong-month.jsonl'), 'wrong month\n')
+    writeManagedSession(join('scratch', 'rollout-too-shallow.jsonl'), 'too shallow\n')
+    writeManagedSession(join('2026', '05', '26', 'nested', 'rollout-too-deep.jsonl'), 'too deep\n')
+
+    const summary = await backfillManagedCodexSessionsIntoSystemHome(
+      resolveCodexSessionBackfillPaths()
+    )
+
+    expect(summary).toMatchObject({
+      scannedFiles: 5,
+      linkedFiles: 1,
+      skippedUnexpectedFiles: 4,
+      failedFiles: 0
+    })
+    expect(
+      existsSync(join(getSystemSessionsRoot(), '2026', '05', '26', 'rollout-valid ü.jsonl'))
+    ).toBe(true)
+    expect(
+      existsSync(join(getSystemSessionsRoot(), '2026', '05', '26', 'session-index.jsonl'))
+    ).toBe(false)
+    expect(existsSync(join(getSystemSessionsRoot(), 'scratch'))).toBe(false)
+  })
+
+  it('never overwrites an existing target file, even with different contents', async () => {
+    writeManagedSession(join('2026', '05', '26', 'rollout-a.jsonl'), 'managed contents\n')
+    const collidingPath = join(getSystemSessionsRoot(), '2026', '05', '26', 'rollout-a.jsonl')
+    mkdirSync(dirname(collidingPath), { recursive: true })
+    writeFileSync(collidingPath, 'user contents\n', 'utf-8')
+
+    const summary = await backfillManagedCodexSessionsIntoSystemHome(
+      resolveCodexSessionBackfillPaths()
+    )
+
+    expect(summary).toMatchObject({ scannedFiles: 1, linkedFiles: 0, skippedExistingFiles: 1 })
+    expect(readFileSync(collidingPath, 'utf-8')).toBe('user contents\n')
+    expect(readAuditActions()).toEqual(['existing', 'run-summary'])
+  })
+
+  it('treats an archived rollout as a durable tombstone across repeated runs', async () => {
+    const relativePath = join(
+      '2026',
+      '05',
+      '26',
+      'rollout-2026-05-26T10-00-00-019f0000-1111-7222-8333-000000000001.jsonl'
+    )
+    writeManagedSession(relativePath, 'managed contents\n')
+    const archivedPath = join(getSystemArchivedSessionsRoot(), relativePath.split(/[/\\]/).at(-1)!)
+    mkdirSync(dirname(archivedPath), { recursive: true })
+    writeFileSync(archivedPath, 'archived contents\n', 'utf-8')
+    const paths = resolveCodexSessionBackfillPaths()
+
+    const first = await backfillManagedCodexSessionsIntoSystemHome(paths)
+    const second = await backfillManagedCodexSessionsIntoSystemHome(paths)
+
+    expect(first).toMatchObject({ linkedFiles: 0, skippedExistingFiles: 1, failedFiles: 0 })
+    expect(second).toMatchObject({ linkedFiles: 0, skippedExistingFiles: 1, failedFiles: 0 })
+    expect(existsSync(join(getSystemSessionsRoot(), relativePath))).toBe(false)
+    expect(readFileSync(archivedPath, 'utf-8')).toBe('archived contents\n')
+    expect(readAuditActions()).toEqual(['run-summary'])
+  })
+
+  it('fails closed when the archived rollout cannot be inspected', async () => {
+    const relativePath = join(
+      '2026',
+      '05',
+      '26',
+      'rollout-2026-05-26T10-00-00-019f0000-1111-7222-8333-000000000005.jsonl'
+    )
+    writeManagedSession(relativePath, 'managed contents\n')
+    fsMockState.failLstatPath = join(
+      getSystemArchivedSessionsRoot(),
+      relativePath.split(/[/\\]/).at(-1)!
+    )
+
+    const summary = await backfillManagedCodexSessionsIntoSystemHome(
+      resolveCodexSessionBackfillPaths()
+    )
+
+    expect(summary).toMatchObject({ linkedFiles: 0, failedFiles: 1 })
+    expect(existsSync(join(getSystemSessionsRoot(), relativePath))).toBe(false)
+    expect(readAuditActions()).toEqual(['failed', 'run-summary'])
+  })
+
+  it('rolls back its new active link when an archive wins the publication race', async () => {
+    const relativePath = join(
+      '2026',
+      '05',
+      '26',
+      'rollout-2026-05-26T10-00-00-019f0000-1111-7222-8333-000000000003.jsonl'
+    )
+    writeManagedSession(relativePath, 'managed contents\n')
+    const archivedPath = join(getSystemArchivedSessionsRoot(), relativePath.split(/[/\\]/).at(-1)!)
+    fsMockState.archivePathAfterLink = archivedPath
+
+    const summary = await backfillManagedCodexSessionsIntoSystemHome(
+      resolveCodexSessionBackfillPaths()
+    )
+
+    expect(summary).toMatchObject({ linkedFiles: 0, skippedExistingFiles: 1, failedFiles: 0 })
+    expect(existsSync(join(getSystemSessionsRoot(), relativePath))).toBe(false)
+    expect(existsSync(archivedPath)).toBe(true)
+    expect(readAuditActions()).toEqual(['run-summary'])
+  })
+
+  it('rolls back its new active link when the archive race probe fails', async () => {
+    const relativePath = join(
+      '2026',
+      '05',
+      '26',
+      'rollout-2026-05-26T10-00-00-019f0000-1111-7222-8333-000000000006.jsonl'
+    )
+    writeManagedSession(relativePath, 'managed contents\n')
+    fsMockState.failLstatPathAfterLink = join(
+      getSystemArchivedSessionsRoot(),
+      relativePath.split(/[/\\]/).at(-1)!
+    )
+
+    const summary = await backfillManagedCodexSessionsIntoSystemHome(
+      resolveCodexSessionBackfillPaths()
+    )
+
+    expect(summary).toMatchObject({ linkedFiles: 0, failedFiles: 1 })
+    expect(existsSync(join(getSystemSessionsRoot(), relativePath))).toBe(false)
+    expect(readAuditActions()).toEqual(['failed', 'run-summary'])
+  })
+
+  it('removes an active duplicate when the same rollout is archived', async () => {
+    const relativePath = join(
+      '2026',
+      '05',
+      '26',
+      'rollout-2026-05-26T10-00-00-019f0000-1111-7222-8333-000000000002.jsonl'
+    )
+    const managedPath = writeManagedSession(relativePath, 'managed contents\n')
+    const activePath = join(getSystemSessionsRoot(), relativePath)
+    const archivedPath = join(getSystemArchivedSessionsRoot(), relativePath.split(/[/\\]/).at(-1)!)
+    mkdirSync(dirname(activePath), { recursive: true })
+    mkdirSync(dirname(archivedPath), { recursive: true })
+    linkSync(managedPath, activePath)
+    linkSync(managedPath, archivedPath)
+
+    const summary = await backfillManagedCodexSessionsIntoSystemHome(
+      resolveCodexSessionBackfillPaths()
+    )
+
+    expect(summary).toMatchObject({ linkedFiles: 0, skippedExistingFiles: 1, failedFiles: 0 })
+    expect(existsSync(activePath)).toBe(false)
+    expect(lstatSync(managedPath).ino).toBe(lstatSync(archivedPath).ino)
+    expect(readAuditActions()).toEqual(['run-summary'])
+  })
+
+  it('isolates an archived active-link removal failure and continues later files', async () => {
+    const archivedRelativePath = join('2026', '05', '26', 'rollout-a.jsonl')
+    const managedPath = writeManagedSession(archivedRelativePath, 'archived contents\n')
+    const activePath = join(getSystemSessionsRoot(), archivedRelativePath)
+    const archivedPath = join(getSystemArchivedSessionsRoot(), 'rollout-a.jsonl')
+    mkdirSync(dirname(activePath), { recursive: true })
+    mkdirSync(dirname(archivedPath), { recursive: true })
+    linkSync(managedPath, activePath)
+    linkSync(managedPath, archivedPath)
+    fsMockState.failUnlinkPath = activePath
+
+    const laterRelativePath = join('2026', '05', '26', 'rollout-b.jsonl')
+    writeManagedSession(laterRelativePath, 'later contents\n')
+
+    const summary = await backfillManagedCodexSessionsIntoSystemHome(
+      resolveCodexSessionBackfillPaths()
+    )
+
+    expect(summary).toMatchObject({
+      scannedFiles: 2,
+      linkedFiles: 1,
+      skippedExistingFiles: 0,
+      failedFiles: 1
+    })
+    expect(existsSync(activePath)).toBe(true)
+    expect(existsSync(join(getSystemSessionsRoot(), laterRelativePath))).toBe(true)
+    expect(readAuditActions().toSorted()).toEqual(['failed', 'hardlink', 'run-summary'].toSorted())
+  })
+
+  it('preserves an ambiguous active file when it differs from the archived rollout', async () => {
+    const relativePath = join(
+      '2026',
+      '05',
+      '26',
+      'rollout-2026-05-26T10-00-00-019f0000-1111-7222-8333-000000000004.jsonl'
+    )
+    writeManagedSession(relativePath, 'managed contents\n')
+    const activePath = join(getSystemSessionsRoot(), relativePath)
+    const archivedPath = join(getSystemArchivedSessionsRoot(), relativePath.split(/[/\\]/).at(-1)!)
+    mkdirSync(dirname(activePath), { recursive: true })
+    mkdirSync(dirname(archivedPath), { recursive: true })
+    writeFileSync(activePath, 'different active contents\n', 'utf-8')
+    writeFileSync(archivedPath, 'archived contents\n', 'utf-8')
+
+    const summary = await backfillManagedCodexSessionsIntoSystemHome(
+      resolveCodexSessionBackfillPaths()
+    )
+
+    expect(summary).toMatchObject({ linkedFiles: 0, skippedExistingFiles: 1, failedFiles: 0 })
+    expect(readFileSync(activePath, 'utf-8')).toBe('different active contents\n')
+    expect(readFileSync(archivedPath, 'utf-8')).toBe('archived contents\n')
+    expect(readAuditActions()).toEqual(['run-summary'])
+  })
+
   it('enqueues a target that appears after the existence probe', async () => {
     fsMockState.raceTargetIntoExistence = true
     writeManagedSession(join('2026', '05', '26', 'rollout-a.jsonl'), 'managed contents\n')
