+++ src/main/runtime/rpc/methods/orchestration-worker-release.ts
@@ -5,10 +5,8 @@ import { defineMethod, type RpcMethod } from '../core'
 import { requiredString } from '../schemas'
 import {
   archiveSummary,
-  exposeWorkerTerminalResource
-} from './orchestration-worker-release-receipt'
-import {
   completeWorkerTerminalRelease,
+  exposeWorkerTerminalResource,
   type WorkerReleaseReceipt
 } from './orchestration-worker-release-completion'
